chore: deserializes the dns reply message
This commit deserializes a dns message received as a response. It does
so by deserializing each of the sections of the Message - Header,
Question, Answer, Authority and Additional individually for better
separation of concerns.

Some of the fields of these sections are fixed in length while some are
variable. Deserialization is done by reading through the byte stream,
and calculating these field values by translating the relevant bytes
into the right data type required for the field.